+++ tools/scene_editor/view_scene.cpp
@@ -280,14 +280,14 @@ void View_Scene::on_draw()
 									pts[i * 2 + 0] = center + vec3(r * circle[i + 0], 0.f).xzy();
 									pts[i * 2 + 1] = center + vec3(r * circle[i + 1], 0.f).xzy();
 								}
-								draw_data.primitives.emplace_back("LineList"_h, pts.data(), pts.size(), cvec4(127, 0, 255, 255));
+								draw_data.primitives.emplace_back("LineList"_h, pts.data(), (uint)pts.size(), cvec4(127, 0, 255, 255));
 								center.y += nav->height;
 								for (auto i = 0; i < n; i++)
 								{
 									pts[i * 2 + 0] = center + vec3(r * circle[i + 0], 0.f).xzy();
 									pts[i * 2 + 1] = center + vec3(r * circle[i + 1], 0.f).xzy();
 								}
-								draw_data.primitives.emplace_back("LineList"_h, pts.data(), pts.size(), cvec4(127, 0, 255, 255));
+								draw_data.primitives.emplace_back("LineList"_h, pts.data(), (uint)pts.size(), cvec4(127, 0, 255, 255));
 								center = nav->node->g_pos;
 								pts[0] = center + r * vec3(+1.f, 0.f, 0.f);
 								pts[1] = pts[0] + vec3(0.f, nav->height, 0.f);
